Add files via upload
Response to the problem statement added to the repository. EDA and data cleaning was conducted to isolate the date when a car model was sold, along with the price of the car model. Stationarity test was conducted on the car prices after 2021, and upon confirmation of stationarity, SARIMA model was applied to the data, with a prediction of the car prices in the next 3-5 years for each car model.